Finding organisation users is not specific to auditing
This class is only related to auditing when it is used in an auditing
context. Therefore, it should not belong inside the auditing namespace.
